Add modifiers for keys in modal mode
A key bound like

	key:modal;modifier

Behaves the way it always has: the modifier is applied to the key that enters modal mode.

A key bound like

	key:modifier;modal

Now behaves as follows: the modifier has to be pressed with the key that triggers the action while in modal mode.

Of course, the two can also be combined:

	key:modifier1;modal;modifier2

behaves as you’d expect: modal mode is activated when `modal` is pressed along with `modifier2`
and the command is executed if `key` is pressed along with `modifier1` while in modal mode.

combinations followed by `:toggle` also work